requirements.txt
asgiref==3.6.0
defusedxml==0.7.1
diff-match-patch==20200713
Django==4.0.5
django-appconf==1.0.5
django-browser-reload==1.6.0
django-import-export==2.8.0
django-pandas==0.6.6
django-rest-framework==0.1.0
django-sitetree==1.17.2
django-tailwind==3.3.0
djangorestframework==3.13.1
et-xmlfile==1.1.0
fonttools==4.38.0
fpdf2==2.5.7
MarkupPy==1.14
numpy==1.24.0
odfpy==1.4.1
openpyxl==3.0.10
pandas==1.5.2
Pillow==9.3.0
psycopg2==2.9.3
python-dateutil==2.8.2
python-dotenv==0.21.0
pytz==2022.7
PyYAML==6.0
six==1.16.0
sqlparse==0.4.3
svg.path==6.2
tablib==3.3.0
tzdata==2022.7
xlrd==2.0.1
xlwt==1.3.0
